requirements.txt
Flask==2.0.3
onnxruntime==1.12.0
torch==1.10.0
torchvision==0.11.2
numpy==1.22.3
Pillow==8.3.2
Werkzeug==2.0.3
scikit-image==0.19.3
apscheduler
requests
colorama
flask_cors